+++ packages/quick_actions/quick_actions_ios/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.6.0+14
 
-* Refactors `FLTQuickActionsPlugin` classes into multiple components. 
+* Refactors `FLTQuickActionsPlugin` class into multiple components. 
 * Increases unit tests coverage to 100%. 
 
 ## 0.6.0+13